Update patch set 1
Patch Set 1:

It won't be made to work anytime soon, if at all.

Vendor repo wouldn't work, since this is compiled from source and not a proprietary blob. The ideal solution here would be to make it work with the opensource libion, but you can probably put the blob into the Tab 2 vendor repo and it'll just overwrite the source-built one during build (At least, that's my understanding of the system).

Patch-set: 1