Update patch set 3
Patch Set 3: Code-Review+1

Since we're talking about a CM repo, they don't need to fork anything.

These devices are being switched over to use the hardware/ti/omap4 repo, which is the proper home for them. This repo duplicates some parts of that one, and will be always late with the updates (all omap4 devices are being updated to use that repo, making them much easier to maintain).

Lastly, we switched to the open source domx/camera HALs found in that repo, instead of the proprietary blobs in the vendor repo (making the device a bit more future-proof).

Patch-set: 3